+++ ImageSource/src/OIIOReader.cpp
@@ -107,7 +107,10 @@ void OIIOReader::open( TextureInfo* info )
         if( !m_input )
         {
             m_input = OIIO::ImageInput::open( m_filename );
-            OTK_ASSERT_MSG( m_input, ( "Failed to open image file " + m_filename + "." ).c_str() );
+            if( !m_input )
+            {
+                throw std::runtime_error( OIIO::geterror().c_str() );
+            }                
 
             OIIO::ImageSpec spec = m_input->spec();